Separate ServiceTest
